Revert changes in circle drawing, move to particle classes
