+++ src/render/draw_particle.js
@@ -4,30 +4,32 @@ import StencilMode from '../gl/stencil_mode.js';
 import DepthMode from '../gl/depth_mode.js';
 import CullFaceMode from '../gl/cull_face_mode.js';
 import Program from './program.js';
-import {particleUniformValues, particleDefinesValues} from './program/particle_program.js';
+import {circleUniformValues, circleDefinesValues} from './program/circle_program.js';
 import SegmentVector from '../data/segment.js';
 import {OverscaledTileID} from '../source/tile_id.js';
+import ColorMode from '../gl/color_mode.js';
+import ImageSource, { globalTexture } from '../source/image_source.js';
 
 import type Painter from './painter.js';
 import type SourceCache from '../source/source_cache.js';
-import type ParticleStyleLayer from '../style/style_layer/particle_style_layer.js';
-import type ParticleBucket from '../data/bucket/particle_bucket.js';
+import type CircleStyleLayer from '../style/style_layer/circle_style_layer.js';
+import type CircleBucket from '../data/bucket/circle_bucket.js';
 import type ProgramConfiguration from '../data/program_configuration.js';
 import type VertexBuffer from '../gl/vertex_buffer.js';
 import type IndexBuffer from '../gl/index_buffer.js';
 import type {UniformValues} from './uniform_binding.js';
-import type {CircleUniformsType as ParticleUniformsType} from './program/particle_program.js';
+import type {CircleUniformsType} from './program/circle_program.js';
 import type Tile from '../source/tile.js';
 import type {DynamicDefinesType} from './program/program_uniforms.js';
 
-export default drawParticles;
+export default drawCircles;
 
 type TileRenderState = {
     programConfiguration: ProgramConfiguration,
     program: Program<*>,
     layoutVertexBuffer: VertexBuffer,
     indexBuffer: IndexBuffer,
-    uniformValues: UniformValues<ParticleUniformsType>,
+    uniformValues: UniformValues<CircleUniformsType>,
     tile: Tile
 };
 
@@ -37,13 +39,13 @@ type SegmentsTileRenderState = {
     state: TileRenderState
 };
 
-function drawParticles(painter: Painter, sourceCache: SourceCache, layer: ParticleStyleLayer, coords: Array<OverscaledTileID>) {
+function drawCircles(painter: Painter, sourceCache: SourceCache, layer: CircleStyleLayer, coords: Array<OverscaledTileID>) {
     if (painter.renderPass !== 'translucent') return;
 
-    const opacity = layer.paint.get('particle-opacity');
-    const strokeWidth = layer.paint.get('particle-stroke-width');
-    const strokeOpacity = layer.paint.get('particle-stroke-opacity');
-    const sortFeaturesByKey = layer.layout.get('particle-sort-key').constantOr(1) !== undefined;
+    const opacity = layer.paint.get('circle-opacity');
+    const strokeWidth = layer.paint.get('circle-stroke-width');
+    const strokeOpacity = layer.paint.get('circle-stroke-opacity');
+    const sortFeaturesByKey = layer.layout.get('circle-sort-key').constantOr(1) !== undefined;
 
     if (opacity.constantOr(1) === 0 && (strokeWidth.constantOr(1) === 0 || strokeOpacity.constantOr(1) === 0)) {
         return;
@@ -55,34 +57,31 @@ function drawParticles(painter: Painter, sourceCache: SourceCache, layer: Partic
     const depthMode = painter.depthModeForSublayer(0, DepthMode.ReadOnly);
     // Turn off stencil testing to allow circles to be drawn across boundaries,
     // so that large circles are not clipped to tiles
+    const gradientMode = true;
     const stencilMode = StencilMode.disabled;
-    const colorMode = painter.colorModeForRenderPass();
+    const colorMode = gradientMode ? ColorMode.additiveBlended : painter.colorModeForRenderPass();
 
-    const segmentsRenderStates: Array<SegmentsTileRenderState> = [];
+    //const segmentsRenderStates: Array<SegmentsTileRenderState> = [];
 
     for (let i = 0; i < coords.length; i++) {
         const coord = coords[i];
 
         const tile = sourceCache.getTile(coord);
-        const bucket: ?ParticleBucket<*> = (tile.getBucket(layer): any);
+        const bucket: ?CircleBucket<*> = (tile.getBucket(layer): any);
         if (!bucket) continue;
+        bucket.update();
 
         const programConfiguration = bucket.programConfigurations.get(layer.id);
-        const definesValues = particleDefinesValues(layer);
-        const program = painter.useProgram('particle', programConfiguration, ((definesValues: any): DynamicDefinesType[]));
+        const definesValues = circleDefinesValues(layer);
+        if (gradientMode) {
+            definesValues.push("PARTICLE_GRADIENT");
+        }
+        const program = painter.useProgram('circle', programConfiguration, ((definesValues: any): DynamicDefinesType[]));
         const layoutVertexBuffer = bucket.layoutVertexBuffer;
         const indexBuffer = bucket.indexBuffer;
-        const uniformValues = particleUniformValues(painter, coord, tile, layer);
 
-        const state: TileRenderState = {
-            programConfiguration,
-            program,
-            layoutVertexBuffer,
-            indexBuffer,
-            uniformValues,
-            tile
-        };
 
+        /*
         if (sortFeaturesByKey) {
             const oldSegments = bucket.segments.get();
             for (const segment of oldSegments) {
@@ -99,16 +98,58 @@ function drawParticles(painter: Painter, sourceCache: SourceCache, layer: Partic
                 state
             });
         }
+        */
+
+        for (var emitter of bucket.system.emitters) {
+            for (var particle of emitter.particles) {
+                
+                const uniformValues = circleUniformValues(painter, coord, tile, layer, 
+                    emitter.location.x + emitter.zoom * particle.locationOffset.x, 
+                    emitter.location.y + emitter.zoom * particle.locationOffset.y, 
+                    emitter.elevation,
+                    particle.opacity,
+                    particle.scale,
+                    particle.color);
+                const segments = bucket.segments;
+        
+                const isGlobeProjection = painter.transform.projection.name === 'globe';
+                const terrainOptions = {useDepthForOcclusion: !isGlobeProjection};
+
+                if (painter.terrain) painter.terrain.setupElevationDraw(tile, program, terrainOptions);
+        
+                painter.prepareDrawProgram(context, program, tile.tileID.toUnwrapped());
+
+                if (globalTexture) {
+                    context.activeTexture.set(gl.TEXTURE0);
+                    globalTexture.bind(gl.LINEAR, gl.CLAMP_TO_EDGE);
+                }
+        
+                program.draw(context, gl.TRIANGLES, depthMode, stencilMode, colorMode, CullFaceMode.disabled,
+                    uniformValues, layer.id,
+                    layoutVertexBuffer, indexBuffer, segments,
+                    layer.paint, painter.transform.zoom, programConfiguration);
+
+            }
+        }
 
     }
 
+    /*
     if (sortFeaturesByKey) {
         segmentsRenderStates.sort((a, b) => a.sortKey - b.sortKey);
     }
+    */
 
-    const isGlobeProjection = painter.transform.projection.name === 'globe';
-    const terrainOptions = {useDepthForOcclusion: !isGlobeProjection};
+    //const isGlobeProjection = painter.transform.projection.name === 'globe';
+    //const terrainOptions = {useDepthForOcclusion: !isGlobeProjection};
 
+    //console.log(segmentsRenderStates);
+
+    //const isGlobeProjection = painter.transform.projection.name === 'globe';
+    //const terrainOptions = {useDepthForOcclusion: !isGlobeProjection};
+
+
+    /*
     for (const segmentsState of segmentsRenderStates) {
         const {programConfiguration, program, layoutVertexBuffer, indexBuffer, uniformValues, tile} = segmentsState.state;
         const segments = segmentsState.segments;
@@ -117,9 +158,10 @@ function drawParticles(painter: Painter, sourceCache: SourceCache, layer: Partic
 
         painter.prepareDrawProgram(context, program, tile.tileID.toUnwrapped());
 
-        program.draw(context, gl.TRIANGLES, depthMode, stencilMode, colorMode, CullFaceMode.disabled,
+        program.draw(context, gl.TRIANGLES, depthMode, stencilMode, ColorMode.alphaBlended, CullFaceMode.disabled,
             uniformValues, layer.id,
             layoutVertexBuffer, indexBuffer, segments,
             layer.paint, painter.transform.zoom, programConfiguration);
     }
+    */
 }
