+++ src/shaders/circle.vertex.glsl
@@ -4,7 +4,6 @@
 
 #define NUM_SAMPLES_PER_RING 16
 
-uniform vec3 u_pos_offset;
 uniform mat4 u_matrix;
 uniform mat2 u_extrude_scale;
 uniform lowp float u_device_pixel_ratio;
@@ -98,8 +97,6 @@ void main(void) {
     // multiply a_pos by 0.5, since we had it * 2 in order to sneak
     // in extrusion data
     vec2 circle_center = floor(a_pos * 0.5);
-
-    circle_center += u_pos_offset.xy;
 
 #ifdef PROJECTION_GLOBE_VIEW
     // Compute positions on both globe and mercator plane to support transition between the two modes
